config/server: improve default cursor appearance
Previously, if the user did not manually configure their cursor
settings, they would end up with the `default` theme (which, if
available, is pretty boring) at a cursor size of 16.

Now, we attempt to use the cursor-shape protocol (if it is available),
or fall back to detecting the cursor theme and size which are in use by
the host compositor (via the XCURSOR_THEME and XCURSOR_SIZE environment
variables).