server: request server-side decorations
Compositors are supposed to assume that clients default to CSD if they
do not use xdg-decoration. We need to explicitly request SSD from
compositors which support it.

There is no listener to respond to the server's requested mode because
we will not be drawing CSD at all. I have no interest in linking to
libdecor or implementing my own CSD for the one mainstream compositor
(mutter) which fails to support a basic use case.