cmd/config/wrap: add `waywall wrap` subcommand
This new `wrap` subcommand allows users to launch single instances of
Minecraft in waywall quite conveniently. When `waywall wrap` is run, a
new instance of waywall is started and the given command is run in that
environment. It's basically just `waywall run` + `waywall exec` except
the spawned waywall instance doesn't provide some of the normal wall
functionality (e.g. auto-pausing) which may be undesirable for practice,
ranked, or general single instancing.

Some other features, such as parts of the Lua API, are also not
available while waywall is running on wrap mode. I would like to work on
this.

This is currently in a usable state but some quality of life
improvements still need to be made: for example, the window shows up
before the instance is actually spawned. It also doesn't take the size
that the window initially spawns with.

Some code quality improvements are probably in order as well; there's a
lot of code duplication between the `wall` and `wrap` modules and the
config API module is a bit of a mess now.