server: add X11 codepath for fake_input
Unfortunately, the only way to send synthetic inputs using the Wayland
protocol is by giving the target surface input focus. This is a problem
for initially pausing on world load (if needed), since SpeedrunIGT will
count it as the start of the run due to the window being focused.

Having a separate codepath for Xwayland toplevels allows for sending
synthetic input events without giving input focus, since the X11
protocol does not use the same enter/leave mechanism as Wayland.