server: do not track opaque region
I'm not actually sure how much of a performance uplift the opaque
optimization can even give, but it adds extra book keeping to
server_region for very little value.

GLFW does appear to use `wl_surface.set_opaque_region`, but Minecraft
does not seem to set the hint necessary for it to take place. Xwayland
does not use it for anything important. So, this shouldn't impact
anything anyway for the time being.