server: redirect xwayland log
This keeps the console output clean while still allowing Xwayland's
console output to be referenced in the event of a problem. Setting the
name of the log file to have the PID of the main waywall process is
intentional so that the log files of waywall and its Xwayland instance
can be associated with one another.