util: log libwayland version
This information is from the `wayland-version.h` header, so this
actually logs the version of the libwayland headers that waywall was
built against (not the actual version of libwayland which is running).

This is desirable to log because I will be using the new connection
buffer size mechanism added in libwayland 1.23, but not every
distribution will have access to the new version yet.