server: increase default wayland buffer size
The default of 4 kilobytes is low enough to cause clients to get
disconnected when they stop reading events for even a short while (e.g.
during first world load lag). 64 kilobytes is still a fairly reasonable
size, but is enough to largely alleviate the problem.

Unfortunately, this buffer size API is only available in the newest
release of libwayland (version 1.23), so it needs to be gated off at
compile time.