server: use "tree" subsurface for layout
Previously, a transaction completing ideally would require that there
were no intermediate calls to `wl_surface.commit` on the root surface.
Now, the transaction state is synchronized by a "tree" subsurface
instead of the root surface.

Additionally, it is not necessary to individually call `set_sync` and
`set_desync` on every view subsurface in a transaction now. It can be
called on the tree subsurface and nothing else.