util: add zip reader
Because we do not need to extract anything from ZIP files, it is easy
enough to implement a basic ZIP reader which is only capable of reading
file metadata. This eliminates the need to have libzip as a dependency.

I still need to implement support for ZIP64, but it seems like all
relevant mods (at least on 1.16.1) do not use it at all. I would hope
nothing does, considering most mods are pretty small, but we'll see.