config: add get_active_res function to lua API
I plan to add a new helper for toggling resolutions easily. Being able
to get the current resolution of the active instance will make it much
more straightforward to implement and use.

I considered making a function to get the resolution of any instance,
but that information is currently not tracked and it would be a very
small value add, I think.